refactor(tooling): group identity contract modules
