+++ scripts/production-infra-identity-contract/workflow-inventory.mjs
