+++ scripts/tf-stacks.test.mjs
@@ -12,7 +12,7 @@ import {
 import { tmpdir } from "node:os";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
-import "./production-infra-identity-contract.test.mjs";
+import "./production-infra-identity-contract/index.test.mjs";
 
 const repoRoot = path.resolve(
   path.dirname(fileURLToPath(import.meta.url)),
